refactor: Move RequestHeaders into its own module
