+++ launchdarkly-server-sdk/src/fdv2/mod.rs
@@ -1,6 +1,7 @@
 pub mod model;
 mod polling;
 mod protocol;
+mod request_headers;
 mod source;
 mod url;
 mod wire;
