+++ launchdarkly-server-sdk/src/fdv2/polling.rs
@@ -8,9 +8,10 @@ use serde::Deserialize;
 
 use super::model::{ChangeSetKind, Selector};
 use super::protocol::{FDv2ProtocolHandler, ProtocolError, ProtocolResult};
+use super::request_headers::RequestHeaders;
 use super::source::{
     read_fallback_directive, ErrorInfo, ErrorKind, FDv1FallbackDirective, FDv2SourceEvent,
-    FDv2SourceResult, RequestHeaders,
+    FDv2SourceResult,
 };
 use super::url::build_fdv2_url;
 use crate::reqwest::is_http_error_recoverable;
